requirements.txt
Django==2.2.13
django-heroku==0.3.1
gunicorn==19.9.0
Jinja2==2.11.2
python-dateutil==2.7.0
whitenoise==4.1
newsapi-python==0.2.6
wit==6.0.0